feat(channels): add filterChannels - closes #44
