+++ lib/channels/index.ts
@@ -1,2 +1,3 @@
 export { default as getChannelsWithReadAbility } from "./getChannelsWithReadAbility";
+export { default as filterChannels } from "./filterChannels";
 export { makeOptions } from "./channels.utils";
